Updated to support more domains
The previous script was working on the basis that the domain was (OU).domain.local which in the setup for new labs is incorrect, since these are using (OU).domain.exclaimersupport.com. This new command does an additional check to find the completed distinguished name and then uses that with the rest of the command script.

Credit to Mr.Norris for some of the syntax here as without him this would've been a lot messier looking to split the results of this out.